+++ ECHO/Core/Classes/SocketOperation/FullAccountSocketOperation.swift
@@ -70,8 +70,8 @@ struct FullAccountSocketOperation: SocketOperation {
         }
     }
     
-    func forceEnd() {
-        let result = Result<[String: UserAccount], ECHOError>(error: ECHOError.connectionLost)
+    func forceEnd(error: ECHOError) {
+        let result = Result<[String: UserAccount], ECHOError>(error: error)
         completion(result)
     }
 }