+++ ECHO/Core/Classes/SocketOperation/GetContractLogsSocketOperation.swift
@@ -60,8 +60,8 @@ struct GetContractLogsSocketOperation: SocketOperation {
         }
     }
     
-    func forceEnd() {
-        let result = Result<[ContractLogEnum], ECHOError>(error: ECHOError.connectionLost)
+    func forceEnd(error: ECHOError) {
+        let result = Result<[ContractLogEnum], ECHOError>(error: error)
         completion(result)
     }
 }