+++ ECHO/Core/Classes/SocketOperation/GetKeyReferencesSocketOperation.swift
@@ -52,8 +52,8 @@ struct GetKeyReferencesSocketOperation: SocketOperation {
         }
     }
     
-    func forceEnd() {
-        let result = Result<[[String]], ECHOError>(error: ECHOError.connectionLost)
+    func forceEnd(error: ECHOError) {
+        let result = Result<[[String]], ECHOError>(error: error)
         completion(result)
     }
 }