+++ ECHO/Core/Classes/SocketOperation/TransactionWithCallbackSocketOperation.swift
@@ -47,4 +47,9 @@ struct TransactionWithCallbackSocketOperation: SocketOperation {
             }
         }
     }
+    
+    func forceEnd(error: ECHOError) {
+        let result = Result<Bool, ECHOError>(error: error)
+        completion(result)
+    }
 }